make json codec
