+++ src/api/http/convert.rs
@@ -1,116 +1,17 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use arrow::array::{Array, Float32Array, Float64Array, StringArray};
+use arrow::array::Array;
 use arrow::datatypes::{DataType, Field, Schema};
 use arrow::record_batch::RecordBatch;
 use serde::Deserialize;
 use serde_json::{Map, Value};
 
+use super::json::{downcast_array, JsonCodec};
 use crate::core::{DType, MurrError, TableSchema};
 
 /// Newtype to implement From<&RecordBatch> (orphan rule prevents impl for serde_json::Value).
 pub struct FetchResponse(pub Value);
-
-trait JsonCodec {
-    type Array: Array + 'static;
-    fn to_json(array: &Self::Array) -> Vec<Value>;
-    fn from_json(values: &[Value]) -> Result<Self::Array, MurrError>;
-}
-
-impl JsonCodec for f32 {
-    type Array = Float32Array;
-
-    fn to_json(array: &Float32Array) -> Vec<Value> {
-        (0..array.len())
-            .map(|i| {
-                if array.is_null(i) {
-                    Value::Null
-                } else {
-                    Value::from(array.value(i))
-                }
-            })
-            .collect()
-    }
-
-    fn from_json(values: &[Value]) -> Result<Float32Array, MurrError> {
-        values
-            .iter()
-            .map(|v| match v {
-                Value::Null => Ok(None),
-                Value::Number(n) => n
-                    .as_f64()
-                    .map(|f| Some(f as f32))
-                    .ok_or_else(|| MurrError::TableError(format!("expected number, got {v}"))),
-                _ => Err(MurrError::TableError(format!("expected number, got {v}"))),
-            })
-            .collect()
-    }
-}
-
-impl JsonCodec for f64 {
-    type Array = Float64Array;
-
-    fn to_json(array: &Float64Array) -> Vec<Value> {
-        (0..array.len())
-            .map(|i| {
-                if array.is_null(i) {
-                    Value::Null
-                } else {
-                    Value::from(array.value(i))
-                }
-            })
-            .collect()
-    }
-
-    fn from_json(values: &[Value]) -> Result<Float64Array, MurrError> {
-        values
-            .iter()
-            .map(|v| match v {
-                Value::Null => Ok(None),
-                Value::Number(n) => n
-                    .as_f64()
-                    .map(Some)
-                    .ok_or_else(|| MurrError::TableError(format!("expected number, got {v}"))),
-                _ => Err(MurrError::TableError(format!("expected number, got {v}"))),
-            })
-            .collect()
-    }
-}
-
-impl JsonCodec for String {
-    type Array = StringArray;
-
-    fn to_json(array: &StringArray) -> Vec<Value> {
-        (0..array.len())
-            .map(|i| {
-                if array.is_null(i) {
-                    Value::Null
-                } else {
-                    Value::String(array.value(i).to_string())
-                }
-            })
-            .collect()
-    }
-
-    fn from_json(values: &[Value]) -> Result<StringArray, MurrError> {
-        values
-            .iter()
-            .map(|v| match v {
-                Value::Null => Ok(None),
-                Value::String(s) => Ok(Some(s.as_str())),
-                _ => Err(MurrError::TableError(format!("expected string, got {v}"))),
-            })
-            .collect()
-    }
-}
-
-fn downcast_array<A: Array + 'static>(array: &dyn Array) -> Result<&A, MurrError> {
-    array
-        .as_any()
-        .downcast_ref::<A>()
-        .ok_or_else(|| MurrError::ArrowError(format!("downcast failed for {:?}", array.data_type())))
-}
 
 impl TryFrom<&RecordBatch> for FetchResponse {
     type Error = MurrError;
@@ -173,6 +74,7 @@ impl WriteRequest {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use arrow::array::{Float32Array, Float64Array, StringArray};
     use crate::core::{ColumnSchema, DType};
 
     fn test_table_schema() -> TableSchema {
